refactor(lib): move Stripper config code out of StripperJob
I want to have two variants of the Stripper job so pulling the core
logic to a separate class.

I think this structure is better anyway, would make it easier to do
Stripper stuff outside of Jobs in future if needed
