--- src/Lumper.Lib/Jobs/StripperJob.Add.cs
@@ -1,25 +1,0 @@
-namespace Lumper.Lib.Jobs;
-
-using System.Collections.Generic;
-using System.IO;
-using Lumper.Lib.Bsp.Lumps.BspLumps;
-using Lumper.Lib.Bsp.Struct;
-using Prop = System.Collections.Generic.KeyValuePair<string, string>;
-
-public partial class StripperJob
-{
-    protected class Add : Block
-    {
-        public List<Prop> Properties { get; set; } = [];
-
-        public override void Parse(StreamReader reader, bool blockOpen, ref int lineNr) =>
-            ParseBlock(reader, blockOpen, ref lineNr, (line, lNr) => Properties.Add(ParseProp(line, lNr)));
-
-        public override void Apply(EntityLump lump)
-        {
-            Entity entity = new(Properties);
-            lump.Data.Add(entity);
-            Logger.Info($"Created entity {entity.PresentableName}");
-        }
-    }
-}
